chore: rename `useOrderProgressBarV2Props` to `useOrderProgressBarProps`
