+++ apps/cowswap-frontend/src/common/hooks/orderProgressBar/atoms.ts
