+++ apps/cowswap-frontend/src/common/hooks/orderProgressBar/index.ts
@@ -1,3 +1,3 @@
-export * from './useOrderProgressBarV2Props'
+export * from './useOrderProgressBarProps'
 export * from './types'
 export * from './ProgressBarV2ExecutingOrdersUpdater'
