+++ apps/cowswap-frontend/src/common/hooks/orderProgressBar/progressBarV2StateMachine.md
