+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/SolvingStep.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { Command } from '@cowprotocol/types'
 
 import { CowSwapAnalyticsCategory, toCowSwapGtmEvent } from 'common/analytics/types'
-import { OrderProgressBarStepName } from 'common/hooks/orderProgressBarV2'
+import { OrderProgressBarStepName } from 'common/hooks/orderProgressBar'
 
 import * as styledEl from './styled'
